action-button: rename component name (darkButton)
the darkbutton component doesn't necessarily mean that it's dark (we
have a light theme), so a better name would be ActionButton
