+++ src/components/index.js
@@ -1,9 +1,9 @@
+export { default as ActionButton } from './ActionButton'
 export { default as ActionList } from './ActionList'
 export { default as AnalysisInfo } from './MessageInfo/Analysis'
 export { default as Autocomplete } from './Autocomplete'
 export { default as Button } from './Button'
 export { default as Cart } from './Cart'
-export { default as DarkButton } from './DarkButton'
 export { default as ErrorBoundary } from './ErrorBoundary'
 export { default as ErrorInfo } from './MessageInfo/Error'
 export { default as Footer } from './Footer'
