+++ stories/ActionButton/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { DarkButton } from '../../src/components'
+import { ActionButton } from '../../src/components'
 import style from './style.css'
 import BoletoIcon from '../../src/images/boleto.svg'
 import CreditCardIcon from '../../src/images/credit-card.svg'
@@ -10,19 +10,19 @@ import CradiCardMoreBoleto from '../../src/images/credit-card-more-boleto.svg'
 
 const onClick = () => null
 
-storiesOf('DarkButton', module)
+storiesOf('ActionButton', module)
   .add('Sample', () => (
     <React.Fragment>
       <div
         className={style.bg}
       >
-        <DarkButton
+        <ActionButton
           icon={<CreditCardIcon />}
           onClick={onClick}
           subtitle="Em até 10x sem juros"
           title="Cartão de crédito"
         />
-        <DarkButton
+        <ActionButton
           icon={<BoletoIcon />}
           onClick={onClick}
           subtitle="Desconto de 25%"
@@ -32,12 +32,12 @@ storiesOf('DarkButton', module)
       <div
         className={style.bg}
       >
-        <DarkButton
+        <ActionButton
           icon={<CradiCardMoreBoleto />}
           onClick={onClick}
           title="Cartão + Boleto"
         />
-        <DarkButton
+        <ActionButton
           icon={<TwoCreditCards />}
           onClick={onClick}
           title="2 cartões"
