+++ stories/index.js
@@ -1,5 +1,5 @@
+import './ActionButton'
 import './Button'
-import './DarkButton'
 import './ErrorBoundary'
 import './Footer'
 import './FormFields'
